+++ inc/streamconnection/IMessage.h
@@ -36,7 +36,7 @@ namespace finalmq {
 
 struct IProtocol;
 
-typedef std::pair<char*, int> BufferRef;
+typedef std::pair<char*, ssize_t> BufferRef;
 
 struct IMessage : public IZeroCopyBuffer
 {
@@ -44,25 +44,25 @@ struct IMessage : public IZeroCopyBuffer
 
     // for send
     virtual void addSendPayload(const std::string& payload) = 0;
-    virtual void addSendPayload(const char* payload, int size) = 0;
-    virtual char* addSendPayload(int size) = 0;
-    virtual void downsizeLastSendPayload(int newSize) = 0;
+    virtual void addSendPayload(const char* payload, ssize_t size) = 0;
+    virtual char* addSendPayload(ssize_t size) = 0;
+    virtual void downsizeLastSendPayload(ssize_t newSize) = 0;
 
     // for receive
     virtual BufferRef getReceivePayload() const = 0;
-    virtual char* resizeReceivePayload(int size) = 0;
+    virtual char* resizeReceivePayload(ssize_t size) = 0;
 
     // for the framework
     virtual const std::list<BufferRef>& getAllSendBuffers() const = 0;
-    virtual int getTotalSendBufferSize() const = 0;
+    virtual ssize_t getTotalSendBufferSize() const = 0;
     virtual const std::list<BufferRef>& getAllSendPayloads() const = 0;
-    virtual int getTotalSendPayloadSize() const = 0;
+    virtual ssize_t getTotalSendPayloadSize() const = 0;
 
     // for the protocol to add a header
     virtual void addSendHeader(const std::string& header) = 0;
-    virtual void addSendHeader(const char* header, int size) = 0;
-    virtual char* addSendHeader(int size) = 0;
-    virtual void downsizeLastSendHeader(int newSize) = 0;
+    virtual void addSendHeader(const char* header, ssize_t size) = 0;
+    virtual char* addSendHeader(ssize_t size) = 0;
+    virtual void downsizeLastSendHeader(ssize_t newSize) = 0;
 
     // for the protocol to prepare the message for send
     virtual void prepareMessageToSend() = 0;
