+++ inc/variant/IVariantValue.h
@@ -22,6 +22,8 @@
 
 #pragma once
 
+#include "helpers/BexDefines.h"
+
 #include <string>
 #include <memory>
 
@@ -35,11 +37,11 @@ class Variant;
 struct IVariantVisitor
 {
     virtual ~IVariantVisitor() {}
-    virtual void enterLeaf(Variant& variant, int type, int index, int level, int size, const std::string& name) = 0;
-    virtual void enterStruct(Variant& variant, int type, int index, int level, int size, const std::string& name) = 0;
-    virtual void exitStruct(Variant& variant, int type, int index, int level, int size, const std::string& name) = 0;
-    virtual void enterList(Variant& variant, int type, int index, int level, int size, const std::string& name) = 0;
-    virtual void exitList(Variant& variant, int type, int index, int level, int size, const std::string& name) = 0;
+    virtual void enterLeaf(Variant& variant, int type, ssize_t index, int level, ssize_t size, const std::string& name) = 0;
+    virtual void enterStruct(Variant& variant, int type, ssize_t index, int level, ssize_t size, const std::string& name) = 0;
+    virtual void exitStruct(Variant& variant, int type, ssize_t index, int level, ssize_t size, const std::string& name) = 0;
+    virtual void enterList(Variant& variant, int type, ssize_t index, int level, ssize_t size, const std::string& name) = 0;
+    virtual void exitList(Variant& variant, int type, ssize_t index, int level, ssize_t size, const std::string& name) = 0;
 };
 
 
@@ -58,8 +60,8 @@ struct IVariantValue
     virtual bool add(const std::string& name, Variant&& variant) = 0;
     virtual bool add(const Variant& variant) = 0;
     virtual bool add(Variant&& variant) = 0;
-    virtual int size() const = 0;
-    virtual void visit(IVariantVisitor& visitor, Variant& variant, int index, int level, int size, const std::string& name) = 0;
+    virtual ssize_t size() const = 0;
+    virtual void visit(IVariantVisitor& visitor, Variant& variant, ssize_t index, int level, ssize_t size, const std::string& name) = 0;
 };
