+++ src/helpers/ZeroCopyBuffer.cpp
@@ -22,6 +22,7 @@
 
 
 
+#include "helpers/BexDefines.h"
 #include "helpers/ZeroCopyBuffer.h"
 
 #include <algorithm>
@@ -34,7 +35,7 @@ namespace finalmq {
 std::string ZeroCopyBuffer::getData() const
 {
     std::string data;
-    int size = 0;
+    ssize_t size = 0;
     std::for_each(m_strings.begin(), m_strings.end(), [&size] (const std::string& entry) {
         size += entry.size();
     });
@@ -49,7 +50,7 @@ std::string ZeroCopyBuffer::getData() const
 
 
 
-char* ZeroCopyBuffer::addBuffer(int size)
+char* ZeroCopyBuffer::addBuffer(ssize_t size)
 {
     assert(size > 0);
     std::string str;
@@ -58,7 +59,7 @@ char* ZeroCopyBuffer::addBuffer(int size)
     return const_cast<char*>(m_strings.back().data());
 }
 
-void ZeroCopyBuffer::downsizeLastBuffer(int newSize)
+void ZeroCopyBuffer::downsizeLastBuffer(ssize_t newSize)
 {
     if (m_strings.empty() && newSize == 0)
     {
