+++ src/json/JsonParser.cpp
@@ -71,7 +71,7 @@ const char* JsonParser::parseWhiteSpace(const char* str)
 }
 
 
-const char* JsonParser::parse(const char* str, int size)
+const char* JsonParser::parse(const char* str, ssize_t size)
 {
     const char* ret = parseValue(str, size);
     m_visitor.finished();
@@ -80,7 +80,7 @@ const char* JsonParser::parse(const char* str, int size)
 
 
 
-const char* JsonParser::parseValue(const char* str, int size)
+const char* JsonParser::parseValue(const char* str, ssize_t size)
 {
     if (size >= CHECK_ON_ZEROTERM)
     {
@@ -260,7 +260,7 @@ const char* JsonParser::parseNumber(const char* str)
         assert(value < 0);
         if (value >= INT_MIN)
         {
-            m_visitor.enterInt32(value);
+            m_visitor.enterInt32(static_cast<std::int32_t>(value));
         }
         else
         {
@@ -277,7 +277,7 @@ const char* JsonParser::parseNumber(const char* str)
         }
         if (value <= INT_MAX)
         {
-            m_visitor.enterUInt32(value);
+            m_visitor.enterUInt32(static_cast<std::uint32_t>(value));
         }
         else
         {
@@ -342,7 +342,7 @@ const char* JsonParser::parseString(const char* str, bool key)
     {
         if (c == '\"')
         {
-            int size = str - strBegin;
+            ssize_t size = str - strBegin;
             if (key)
             {
                 m_visitor.enterKey(strBegin, size);
